List field handlers refactoring
* Created BaseListFieldHandler class for all lists
* Changed DblistFieldHandler to extend BaseListFieldHandler class
* Created BaseCsvListFieldHandler class
* Changed ListFieldHandler to extend BaseCsvListFieldHandler class
* Changed SublistFieldHandler to extend BaseCsvListFieldHandler class
* Moved common settings and functionality into base clases
