+++ src/FieldHandlers/BaseListFieldHandler.php
@@ -1,0 +1,53 @@
+<?php
+namespace CsvMigrations\FieldHandlers;
+
+use CsvMigrations\FieldHandlers\BaseFieldHandler;
+
+/**
+ * BaseListFieldHandler
+ *
+ * This class provides the fallback functionality that
+ * is common to list field handlers.
+ */
+abstract class BaseListFieldHandler extends BaseFieldHandler
+{
+    /**
+     * Search operators
+     *
+     * @var array
+     */
+    public $searchOperators = [
+        'is' => [
+            'label' => 'is',
+            'operator' => 'IN',
+        ],
+        'is_not' => [
+            'label' => 'is not',
+            'operator' => 'NOT IN',
+        ],
+    ];
+
+    /**
+     * Convert CsvField to one or more DbField instances
+     *
+     * Simple fields from migrations CSV map one-to-one to
+     * the database fields.  More complex fields can combine
+     * multiple database fields for a single CSV entry.
+     *
+     * @param  \CsvMigrations\FieldHandlers\CsvField $csvField CsvField instance
+     * @return array                                           DbField instances
+     */
+    public function fieldToDb(CsvField $csvField)
+    {
+        $dbFields[] = new DbField(
+            $csvField->getName(),
+            static::DB_FIELD_TYPE,
+            null,
+            $csvField->getRequired(),
+            $csvField->getNonSearchable(),
+            $csvField->getUnique()
+        );
+
+        return $dbFields;
+    }
+}
