+++ src/FieldHandlers/DblistFieldHandler.php
@@ -1,31 +1,10 @@
 <?php
 namespace CsvMigrations\FieldHandlers;
 
-use CsvMigrations\FieldHandlers\BaseFieldHandler;
+use CsvMigrations\FieldHandlers\BaseListFieldHandler;
 
-class DblistFieldHandler extends BaseFieldHandler
+class DblistFieldHandler extends BaseListFieldHandler
 {
-    /**
-     * Field type
-     */
-    const DB_FIELD_TYPE = 'string';
-
-    /**
-     * Search operators
-     *
-     * @var array
-     */
-    public $searchOperators = [
-        'is' => [
-            'label' => 'is',
-            'operator' => 'IN',
-        ],
-        'is_not' => [
-            'label' => 'is not',
-            'operator' => 'NOT IN',
-        ],
-    ];
-
     /**
      * Input default options
      *
@@ -132,29 +111,5 @@ class DblistFieldHandler extends BaseFieldHandler
         ];
 
         return $result;
-    }
-
-    /**
-     * Convert CsvField to one or more DbField instances
-     *
-     * Simple fields from migrations CSV map one-to-one to
-     * the database fields.  More complex fields can combine
-     * multiple database fields for a single CSV entry.
-     *
-     * @param  \CsvMigrations\FieldHandlers\CsvField $csvField CsvField instance
-     * @return array                                           DbField instances
-     */
-    public function fieldToDb(CsvField $csvField)
-    {
-        $dbFields[] = new DbField(
-            $csvField->getName(),
-            static::DB_FIELD_TYPE,
-            null,
-            $csvField->getRequired(),
-            $csvField->getNonSearchable(),
-            $csvField->getUnique()
-        );
-
-        return $dbFields;
     }
 }
